+++ src/lib/mina/local-blockchain.ts
@@ -38,25 +38,29 @@ import {
   verifyTransactionLimits,
   verifyAccountUpdate,
 } from './transaction-validation.js';
+import { prettifyStacktrace } from '../util/errors.js';
 
 export { LocalBlockchain };
+
 /**
  * A mock Mina blockchain running locally and useful for testing.
  */
 function LocalBlockchain({
   proofsEnabled = true,
   enforceTransactionLimits = true,
-  networkId = 'testnet' as NetworkId,
 } = {}) {
   const slotTime = 3 * 60 * 1000;
   const startTime = Date.now();
   const genesisTimestamp = UInt64.from(startTime);
   const ledger = Ledger.create();
   let networkState = defaultNetworkState();
-  let minaNetworkId: NetworkId = networkId;
 
   function addAccount(publicKey: PublicKey, balance: string) {
-    ledger.addAccount(Ml.fromPublicKey(publicKey), balance);
+    try {
+      ledger.addAccount(Ml.fromPublicKey(publicKey), balance);
+    } catch (error) {
+      throw prettifyStacktrace(error);
+    }
   }
 
   let testAccounts: {
@@ -80,7 +84,7 @@ function LocalBlockchain({
   > = {};
 
   return {
-    getNetworkId: () => minaNetworkId,
+    getNetworkId: () => 'testnet' as NetworkId,
     proofsEnabled,
     getNetworkConstants() {
       return {
@@ -121,7 +125,7 @@ function LocalBlockchain({
       let zkappCommandJson = ZkappCommand.toJSON(txn.transaction);
       let commitments = transactionCommitments(
         TypesBigint.ZkappCommand.fromJSON(zkappCommandJson),
-        minaNetworkId
+        this.getNetworkId()
       );
 
       if (enforceTransactionLimits) verifyTransactionLimits(txn.transaction);
@@ -302,7 +306,7 @@ function LocalBlockchain({
     },
     transaction(sender: FeePayerSpec, f: () => Promise<void>) {
       return toTransactionPromise(async () => {
-        // TODO we run the transaction twice to match the behaviour of `Network.transaction`
+        // TODO we run the transaction twice to match the behavior of `Network.transaction`
         let tx = await createTransaction(sender, f, 0, {
           isFinalRunOutsideCircuit: false,
           proofsEnabled: this.proofsEnabled,